bus-booking/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
httptools==0.6.1

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9          # for Celery sync tasks

# ─── Auth & Security ──────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ─── Redis ────────────────────────────────────────────────────────────────────
redis==5.0.4

# ─── Celery ───────────────────────────────────────────────────────────────────
celery==5.3.6
flower==2.0.1

# ─── Payment ──────────────────────────────────────────────────────────────────
razorpay==1.4.1

# ─── Config & Validation ──────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1

# ─── Rate limiting ────────────────────────────────────────────────────────────
slowapi==0.1.9

# ─── Logging ──────────────────────────────────────────────────────────────────
python-json-logger==2.0.7

# ─── Monitoring ───────────────────────────────────────────────────────────────
sentry-sdk[fastapi]==2.1.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-cov==5.0.0
factory-boy==3.3.0
